requirements.txt
Django==1.7.4
Unipath==1.1
argparse==1.2.1
dj-database-url==0.3.0
django-chartit==0.1
simplejson==3.6.5
wsgiref==0.1.2
